multimodal_ai/requirements.txt
# 核心依赖
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
numpy>=1.21.0

# Web框架
flask>=2.0.0
flask-cors>=3.0.10

# 数据处理
pillow>=8.3.0
opencv-python>=4.5.0
librosa>=0.8.1
scipy>=1.7.0

# 工具库
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0
requests>=2.26.0

# 配置和日志
pyyaml>=5.4.0
colorlog>=6.4.0

# 测试
pytest>=6.2.0
pytest-cov>=2.12.0

# 可选依赖（用于模型导出）
onnx>=1.10.0
onnxruntime>=1.9.0

# 开发工具
black>=21.7.0
flake8>=3.9.0
isort>=5.9.0

# 系统监控
psutil>=5.8.0

# 数据科学
pandas>=1.3.0
scikit-learn>=1.0.0

# 多媒体处理
imageio>=2.9.0
imageio-ffmpeg>=0.4.0

# 文本处理
jieba>=0.42.1
regex>=2021.8.0

# 并行处理
joblib>=1.0.0

# 内存优化
memory-profiler>=0.58.0

# 类型检查
mypy>=0.910
